Updating for Artful Aardvark
It seems that Ubuntu changed how interfaces are named from Xenial Xarus to Artful Aardvark. Instead of VGA1 or HDMI1, it's now VGA-1 or HDMI-1 which threw off the script. 

If you're actually using this, just rewrite the interfaces to whatever shows up as the output for your 'xrandr' command.